begin shrinking it down

+++ src/uci.h
@@ -33,7 +33,7 @@ namespace Stockfish {
 class Position;
 class Move;
 class Score;
-enum Square : int;
+enum Square : int16_t;
 using Value = int;
 
 class UCIEngine {
